+++ db/migrate/20250502210440_create_best50_and100_results.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class CreateBest50And100Results < ActiveRecord::Migration[6.1]
+  def up
+    create_view :best50_and100_results, version: 1
+  end
+
+  def down
+    # Use raw SQL for robust dropping and existence checking
+    execute 'DROP VIEW IF EXISTS best50_and100_results;'
+  end
+end