+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 _Please, add the latest build info on top of the list; use Version::MAJOR only after gold release; keep semantic versioning in line with framework's_
 
+- **0.8.05** [Steve A.] updated all "Best Result" views common abstract base; improved specs; added db:dump_remove_sandbox task; DB vers. 2.08.05; vers. 0.8.05
+- **0.8.04** [Steve A.] added "Best Result" family of views with models and abstract base; DB vers. 2.08.03
 - **0.8.03** [Steve A.] applied patch for ActiveSupport logger; bundle updated
 - **0.8.01** [Steve A.] added Best50mResult Scenic view with dedicated model and specs; DB vers. 2.08.01
 - **0.8.00** [Steve A.] safe navigation in SwimmerDecorator#display_label when gender is not defined; added bias parameter in CmdFindDBEntity <- DbFinders::FuzzySwimmer because sometimes the default 0.8 doesn't fit the case; refactored Swimmer#age_for_category_range from [Admin2]PdfResults::L2Converter; bundle update